Add Report Service and Financial Reporting Features
- Introduced IReportService interface and ReportService class to handle financial and tenant reporting functionalities.
- Implemented methods for generating financial reports and tenant reports based on user ID and date range.
- Created FinancialReportDto and TenantReportDto for structured data representation in reports.
- Added ReportsController to expose endpoints for retrieving financial and tenant reports, ensuring access control for desktop users.
- Updated Program.cs to register the new report services in the dependency injection container.
- Enhanced the desktop application to integrate report providers for fetching and displaying report data.
- Refactored existing components to accommodate new reporting features and improve user experience in the reports section.

+++ eRents.Shared/DTO/Response/PropertyResponseDto.cs
@@ -1,27 +1,27 @@
 namespace eRents.Shared.DTO.Response
 {
-    public class PropertyResponseDto
-    {
-        public int Id { get; set; }
-        public int OwnerId { get; set; }
-        public string Name { get; set; } = null!;
-        public string? Description { get; set; }
-        public decimal Price { get; set; }
-        public string? Currency { get; set; }
-        public string Status { get; set; } = null!;
-        public DateTime DateAdded { get; set; }
-        public int? Bedrooms { get; set; }
-        public int? Bathrooms { get; set; }
-        public double? Area { get; set; }
-        public decimal? DailyRate { get; set; }
-        public int? MinimumStayDays { get; set; }
-        public List<ImageResponseDto> Images { get; set; } = new List<ImageResponseDto>();
-        public AddressDetailResponseDto? AddressDetail { get; set; }
-        public List<string> Amenities { get; set; } = new List<string>();
-        
-        // Display properties
-        public string Title => Name; // Alias for compatibility
-        public decimal? AverageRating { get; set; }
-        public int ReviewCount { get; set; }
-    }
-} 
+	public class PropertyResponseDto
+	{
+		public int Id { get; set; }
+		public int OwnerId { get; set; }
+		public string Name { get; set; } = null!;
+		public string? Description { get; set; }
+		public decimal Price { get; set; }
+		public string? Currency { get; set; }
+		public string Status { get; set; } = null!;
+		public DateTime DateAdded { get; set; }
+		public int? Bedrooms { get; set; }
+		public int? Bathrooms { get; set; }
+		public decimal? Area { get; set; }
+		public decimal? DailyRate { get; set; }
+		public int? MinimumStayDays { get; set; }
+		public List<ImageResponseDto> Images { get; set; } = new List<ImageResponseDto>();
+		public AddressDetailResponseDto? AddressDetail { get; set; }
+		public List<string> Amenities { get; set; } = new List<string>();
+
+		// Display properties
+		public string Title => Name; // Alias for compatibility
+		public decimal? AverageRating { get; set; }
+		public int ReviewCount { get; set; }
+	}
+}
